fix(hgraph): restore reverse edges before removal after reload
Rebuild derived incoming edges for flat and sparse graph storage after deserialization. Cover fresh FP32 snapshots, decoded node versions, tail moves, and subsequent mutation. Clarify existing lock and view lifetime contracts and fix the deletion-count narrowing warning.

+++ src/algorithm/hgraph/hgraph_serialize.cpp
@@ -1092,7 +1092,7 @@ HGraph::Deserialize(StreamReader& reader) {
             if (this->support_force_remove()) {
                 const auto removed_ids = this->mci_cliques_->GetInactiveNodeIds();
                 this->label_table_->RestoreDeletedIds(removed_ids, this->total_count_.load());
-                this->delete_count_.store(removed_ids.size());
+                this->delete_count_.store(static_cast<int64_t>(removed_ids.size()));
             }
         }
         if (metadata->Get("has_conjugate_graph").IsBool() &&
